Generalize calendar maintenance codes, fix bugs
